pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "TacoBot"
authors = [
    {name = "Vincent Lin", email = "vinlin24@outlook.com"},
]
description = "Discord bot project"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
keywords = ["discord", "bot", "discord.py"]
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "autopep8~=2.0.2",
    "boto3~=1.24.17",
    "discord.py~=2.2.2",
    "ffmpeg~=1.4",
    "Pillow~=9.1.1",
    "PubChemPy~=1.0.4",
    "pylint~=2.17.2",
    "PyNaCl~=1.5.0",
    "python-dotenv~=0.20.0",
    "requests~=2.28.0",
    "setuptools~=58.1.0",
    "sympy~=1.10.1",
    "tekore~=4.4.0",
    "yt_dlp~=2023.3.4",
]
dynamic = ["version"]

[tool.setuptools.dynamic]
version = {attr = "bot.__version__"}
